removed exp section from develop branch and removed Main wrapper component
