+++ src/Components/App.jsx
@@ -1,5 +1,10 @@
 import React, { useState, useEffect, createContext } from "react";
-import Main from "./Main";
+import NavBar from "./NavBar";
+import About from "./About";
+import Projects from "./Projects";
+import Skills from "./Skills";
+import Contact from "./Contact";
+import Intro from "./Intro";
 import "../css/app.css";
 
 export const ThemeContext = createContext(null);
@@ -42,10 +47,42 @@ function App() {
     body.style.backgroundColor = "#FFFFFF";
     body.style.color = "#000000";
   }
+
+  // Get all sections that have an ID defined
+	const sections = document.querySelectorAll("section[id]");
+
+	// Add an event listener listening for scroll
+	window.addEventListener("scroll", navHighlighter);
+
+	function navHighlighter() {
+		// Get current scroll position
+		let scrollY = window.scrollY;
+
+		// Now we loop through sections to get height, top and ID values for each
+		sections.forEach(current => {
+			const sectionHeight = current.offsetHeight;
+			const sectionTop = current.offsetTop - 62;
+			var sectionId = current.getAttribute("id");
+			/*
+			- If our current scroll position enters the space where current section on screen is, add .active class to corresponding navigation link, else remove it
+			- To know which link needs an active class, we use sectionId variable we are getting while looping through sections as an selector
+			*/
+			if (scrollY > sectionTop && scrollY <= sectionTop + sectionHeight) {
+				document.querySelector(".my-nav a[href*=" + sectionId + "]").classList.add("active");
+			} else {
+				document.querySelector(".my-nav a[href*=" + sectionId + "]").classList.remove("active");
+			}
+		});
+	}
   return (
     <ThemeContext.Provider value={{themeflag, theme, toggleTheme}}>
       <div className="app" id={theme}>
-        <Main />
+        <NavBar />
+        <Intro />
+        <About />
+        <Projects />
+        <Skills />
+        <Contact />
 	    </div>
     </ThemeContext.Provider>
   );
