--- src/css/experience.css
@@ -1,38 +1,0 @@
-.experience {
-	width: 100%;
-	min-height: 100%;
-
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	flex-direction: column;
-
-    padding-top: 52px;
-    /* margin-top: -52px; margin removed due to nav highlighting issue */
-}
-.experience-description {
-	margin: 0 auto;
-    width: auto;
-	/* max-width: 90%; */
-	position: relative;
-	top: 5%;
-}
-.open-source h3 {
-	display: inline;
-}
-.link { text-decoration: none; }
-.link:hover { text-decoration:underline; }
-.github-icon { 
-    position: relative;
-    bottom: 3px;
-}
-@media only screen and (max-width: 576px) {
-	.open-source h3 {
-		display: block;
-	}
-}
-@media only screen and (max-width: 769px) {
-	.experience-description {
-		max-width: 90%;
-	}
-}
