--- src/css/main.css
@@ -1,12 +1,0 @@
-* {
-	font-family: 'Open Sans', sans-serif;
-	box-sizing: border-box;
-}
-html, body, #root {
-	width: 100%;
-	height: 100%;
-}
-.main {
-	width: 100%;
-	height: 100%;
-}
